Refactor: 카드 아이템 내부로 모달창 이동
