+++ src/app/community/_components/PostCard.tsx
@@ -1,20 +1,33 @@
+'use client';
+
+import { useState } from 'react';
 import classNames from 'classnames/bind';
-import ContentImage from '@/public/images/myProfile.jpeg';
 import Image from 'next/image';
+
+import ContentImage from '@/public/images/myProfile.jpeg';
 import { CommunityCardDataType } from '@/app/mj/CommunityData';
 import { calculateTimeDifference } from '@/libs/calculateDate';
-import styles from './PostCard.module.scss';
+
+import { Modal } from '@/components';
 import AuthorCard from './AuthorCard';
 import { PostInteractions } from './PostInteractions';
+
+import styles from './PostCard.module.scss';
 
 const cn = classNames.bind(styles);
 
 interface PostCardProps {
   cardData: CommunityCardDataType;
-  onClick: () => void;
 }
 
-export default function PostCard({ cardData, onClick }: PostCardProps) {
+export default function PostCard({ cardData }: PostCardProps) {
+  const [isPostModalOpen, setIsPostModalOpen] = useState(false);
+  const handleClickPostModal = () => {
+    setIsPostModalOpen(true);
+  };
+  const handleClosePostModal = () => {
+    setIsPostModalOpen(false);
+  };
   const {
     user_nickname: nickname,
     created_at: createdAt,
@@ -25,18 +38,20 @@ export default function PostCard({ cardData, onClick }: PostCardProps) {
   } = cardData;
   const nowDate = new Date();
   const createdDate = new Date(createdAt);
-
   const timeToString = calculateTimeDifference(createdDate, nowDate);
 
   return (
-    <div className={cn('container')} onClick={onClick}>
-      <AuthorCard nickname={nickname} timeAgoString={timeToString} />
+    <div className={cn('container')} onClick={handleClickPostModal}>
+      <AuthorCard nickname={nickname} timeAgo={timeToString} />
       <div className={cn('keyboard-image-wrapper')}>
         <Image src={ContentImage} className={cn('keyboard-image')} alt='키보드 이미지' />
         {image.length > 1 && <p id={cn('image-count')}>{image.length}</p>}
       </div>
       <p className={cn('title')}>{title}</p>
       <PostInteractions goodCount={goodCount} commentCount={commentCount} />
+      <Modal isOpen={isPostModalOpen} onClose={handleClosePostModal}>
+        modal
+      </Modal>
     </div>
   );
 }
